+++ .storybook/preview.tsx
@@ -67,7 +67,7 @@ const preview: Preview = {
         dark: 'dark',
       },
       defaultTheme: 'light',
-      parentSelector: 'body',
+      parentSelector: 'html',
     }),
   ],
 };