+++ src/surveys/NPS10Survey/NPS10Survey.tsx
@@ -3,11 +3,11 @@ import React from 'react';
 import { SurveyRoot } from '../../components/SurveyRoot';
 import useSurveyState from '../../hooks/useSurveyState';
 import type { SharedSurveyProps } from '../../types';
+import { cn } from '../../utils';
 
 import { NPS10SurveyNumbers } from './NPS10SurveyNumbers';
 
 import styles from './NPS10Survey.module.scss';
-import { cn } from '../../utils';
 
 export interface NPS10SurveyProps extends SharedSurveyProps {
   /** Visual style for the rating scale */
@@ -26,17 +26,28 @@ export const NPS10Survey: React.FC<NPS10SurveyProps> = ({
   responseType,
   choiceOptions,
   thankYouMessage,
+  collectContact,
+  userId,
+  contactQuestion,
+  contactSubtext,
+  contactButtonSendLabel,
+  contactButtonSkipLabel,
   onScoreSubmit,
-  onFeedbackSubmit
+  onFeedbackSubmit,
+  onContactSubmit
 }) => {
   const {
     screen,
     onScoreChange,
-    onFeedbackChange
+    onFeedbackChange,
+    onContactChange
   } = useSurveyState({
     responseType,
+    collectContact,
+    userId,
     onScoreSubmit,
-    onFeedbackSubmit
+    onFeedbackSubmit,
+    onContactSubmit
   })
 
   return (
@@ -46,6 +57,7 @@ export const NPS10Survey: React.FC<NPS10SurveyProps> = ({
         ...classNames?.base,
         rating: cn(styles.rating, classNames?.base?.rating),
         feedback: cn(styles.feedback, classNames?.base?.feedback),
+        contact: cn(styles.contact, classNames?.base?.contact),
         success: cn(styles.success, classNames?.base?.success)
       }}
       dir={dir}
@@ -56,7 +68,12 @@ export const NPS10Survey: React.FC<NPS10SurveyProps> = ({
       responseType={responseType}
       choiceOptions={choiceOptions}
       thankYouMessage={thankYouMessage}
+      contactQuestion={contactQuestion}
+      contactSubtext={contactSubtext}
+      contactButtonSendLabel={contactButtonSendLabel}
+      contactButtonSkipLabel={contactButtonSkipLabel}
       onFeedback={onFeedbackChange}
+      onContact={onContactChange}
     >
       {(scaleStyle === 'numbers') && (
         <NPS10SurveyNumbers